coe: put code in `DepTypeTerm`
